Integrates the 23 Phase-4 commits that landed on master since this branch's base
(native reflection bridge, cross-package //go:linkname forwarder, native
atomic.Value / bytealg, reflect.Type canonicalization, untyped-int->nint interface
boxing, switch-fallthrough default guard) with the golib @string/sstring
performance initiative (zero-alloc u8 concat + comparison operators, the sstring
stack-string promotion, and the escape-analysis-driven sstring emission MVP).

Clean auto-merge: master's converter changes (nint boxing, variadic-any interface
fan-out) sit in different regions of convCallExpr.go / visitAssignStmt.go than the
sstring retarget/emission hooks, so no manual conflict resolution was needed.

Verified as a superset of master:
- merged converter builds; sstring emission survives (SStringElision golden byte-identical)
- check-no-regression: zero converter drift across 383 behavioral projects
- merged golib compiles; full go-src-converted.slnx = 302/302 packages, 0 errors
- behavioral transpile+compile green against the merged golib

+++ src/Tests/Behavioral/FuncLitStringConcatReturn/main.cs
@@ -22,7 +22,7 @@ internal static void Main() {
         }}
     };
     fmt.Println(pick("abc"));
-    fmt.Println(pick(42));
+    fmt.Println(pick((nint)(42)));
     fmt.Println(pick(true));
     var wrap = @string (@string s, bool quote) => {
         if (quote) {
